Adding a new example project
current example project is missing important files (like main.m or ViewController.h). Creating a brand new example project that should just run out of the box. This should not be hooked up to the SDKGenerator at all, so this should be a standalone project that should just run right out of the box.