Adding a new example project (#16)
* Adding a new example project

current example project is missing important files (like main.m or ViewController.h). Creating a brand new example project that should just run out of the box. This should not be hooked up to the SDKGenerator at all, so this should be a standalone project that should just run right out of the box.

* Update .gitignore

* Update .gitignore

* delete mac desktop config files

automatically added by mac :(

* removing personal UI config files

added them to .gitignore file as well